Implement dedicated ListModel for handling YouTube list responses.
Currently parsed json data is feed into plain qml ListModel from
javascrtipt code. This has a bad effect for animation smoothnes.
The solution is to use a custom ListModel implementation that can be
populated dynamically from C++ code. The new YTListModel can be
associated with a YTRequest. When such connection is made the request
will automatically append newly fetched data to the model.