Implement support for downloading videos to device storage
This feature could be also called preloading. It allows the user to
download video data directly to device storage. Downloaded videos can be
played without making any request to YouTube servers. Due to the lack of
any download manager linke funcionality in public sailfish APIs the
entire download logic is part of the app. This means the application
needs to be running for the whole duration of the download process.