HeaderButton: Align the look and feel with Sailfish header buttons.
HeaderButton is not exactly a standard component. It's basically a
button for attached pages. In Sailfish similar approach is used for
dialog buttons. In case of YTPlayer attached pages are used for video
player and advanced search settings. Those use cases don't fall under
dialog category. Still It's good to have some  indication what the
attached page contains. HeadersButton is the component supposed to help
with this use caes. Unfortunately the original design does not look that
good on Update 10. This patch tries to fix it.