Add video duration indicator to various video lists
Unfortunately adding duration indicator field for all video lists is not
feasible. YouTube data API v3 does provide duration for video listings,
but not for playlist or activity lists. Doing that for those cases would
require an extra network request along with code and infrastructure to
assign duration responses from second request to data obtained in the
first one. Doable, but due to the comlexity of this operation I'd rather
not do it for 0.4.x release.