VideoOverview: Push ChannelBrowser page instead of replacing VideoOverview
After using the version where channel browser page would replace video
overview when "Browse channel" menu option was selected I decided it was
not very intuitive. The main reason for previous version was to avoid
having to many levels of attached pages as this would add a lot of extra
page dot indicators. In Update 10 this is no longer a problem.