Add common code for network session handling.
In Update 10 I started seeing NetworkSessionFailedError(s) being sent
from QNetworkAccessManager after switching from cellular network to
WiFi. This does not happen once or twice after the switch. The state in
which the error is generated for each request persists for as long as
QNetworkAccessManager::setConfiguration is not called with a new valid
config. Previously I've added some code doing it to YTRequest's QNAM,
but this did not solve the problem for image and video download
requests. To solve the issue allow YTNetworkManager to manage network
session and configs for all instances of QNetworkAccessManager used by
the app.