Implement support for obtaining video streams via youtube-dl application
Since ytapi.com is no longer usable an alternative solution needed to be
found. Youtube-dl seems to be a reasonable alternative. It's a command line
application that can be used to download videos from YouTube servers.
Besides downloading videos it also makes it possible to only obtain list
of available stream urls in json format. It's written in python which is
not ideal for C++ app like YTPlayer, but it is usable. SailfishOS does ship
python interpreter and all the necessary deps to run it.

This patch implements initial support for using youtube-dl as video url
fetcher. This is enough to play almost every video I've found, including
VEVO ones. Performance is not ideal, but it should be possible to
improve it with some additional caching.