Implement additional checks for properly functioning youtube-dl script.
I'm not sure how official python support on Sailfish is so it's better
to be safe. In case the script does not work avoid calling it for each
video which requires extra work to get stream for. Such videos should
not display Play button in the top right corner of VideoOverview page.